refactor: share text AST node loader
